# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API Title",
        "description": "Your API Description",
        "contact": {
            "email": "contact@example.com"
        },
        "license": {
            "name": "License Name",
            "url": "http://www.example.com/license"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/exercise/done": {
            "get": {
                "summary": "Retrieve done exercises with pagination 15 items per page",
                "operationId": "1d744e85ff3aeb54d360c7003989a859",
                "responses": {
                    "200": {
                        "description": "Done exercises retrieved successfully"
                    }
                }
            },
            "post": {
                "summary": "Create or update a done exercise",
                "operationId": "020f65c0639deee22bb9757528d5f1f4",
                "parameters": [
                    {
                        "name": "exercise_id",
                        "in": "query",
                        "description": "The ID of the exercise",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "reps",
                        "in": "query",
                        "description": "The number of repetitions of the exercise",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "kg",
                        "in": "query",
                        "description": "The weight used in the exercise, in kilograms",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "rir",
                        "in": "query",
                        "description": "The reps in reserve",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "tempo",
                        "in": "query",
                        "description": "The tempo of the exercise",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "rest",
                        "in": "query",
                        "description": "The rest time in seconds",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "The status of the exercise",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Done exercise created or updated successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/exercise/createWithPlan": {
            "post": {
                "tags": [
                    "Exercise"
                ],
                "summary": "Create an exercise with a plan",
                "description": "Create a new exercise entry associated with a plan",
                "operationId": "createWithPlan",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "exercise_name": {
                                        "type": "string",
                                        "example": "Push Up"
                                    },
                                    "duration": {
                                        "type": "integer",
                                        "example": 30
                                    },
                                    "plan_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Exercise created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Exercise created successfully"
                                        },
                                        "exercise": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "exercise_name": {
                                                    "type": "string",
                                                    "example": "Push Up"
                                                },
                                                "duration": {
                                                    "type": "integer",
                                                    "example": 30
                                                },
                                                "plan_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/exercise/done/{id}": {
            "get": {
                "summary": "Retrieve a specific done exercise",
                "operationId": "70c48ef3ab2dfdb3d5c9f97337486c9b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Done exercise's id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Done exercise retrieved successfully"
                    },
                    "404": {
                        "description": "Done exercise not found"
                    }
                }
            },
            "put": {
                "summary": "Update a specific done exercise",
                "operationId": "a6eda84d5a73d9ada11354ab32eac08f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Done exercise's id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "exercise_id",
                        "in": "query",
                        "description": "Exercise's id",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "plan_id",
                        "in": "query",
                        "description": "Plan's id",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "reps",
                        "in": "query",
                        "description": "Reps",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "kg",
                        "in": "query",
                        "description": "Kg",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "rir",
                        "in": "query",
                        "description": "Rir",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "tempo",
                        "in": "query",
                        "description": "Tempo",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "rest",
                        "in": "query",
                        "description": "Rest",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Done exercise updated successfully"
                    },
                    "404": {
                        "description": "Done exercise not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete a specific done exercise",
                "operationId": "f585512c96297413c6e3aa2953f9a20e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Done exercise's id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Done exercise deleted successfully"
                    },
                    "404": {
                        "description": "Done exercise not found"
                    }
                }
            }
        },
...
    },
...
}